+++ internal/util/filter/filter_test.go
@@ -145,3 +145,95 @@ func TestNot(t *testing.T) {
 		})
 	}
 }
+
+func TestFilter(t *testing.T) {
+	tests := []struct {
+		name      string
+		slice     []int
+		predicate filter.Predicate[int]
+		want      []int
+	}{
+		{
+			name:      "all match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 0 },
+			want:      []int{1, 2, 3, 4, 5},
+		},
+		{
+			name:      "some match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 3 },
+			want:      []int{4, 5},
+		},
+		{
+			name:      "none match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 5 },
+			want:      nil,
+		},
+		{
+			name:      "empty slice",
+			slice:     []int{},
+			predicate: func(i int) bool { return i > 5 },
+			want:      nil,
+		},
+		{
+			name:      "nil slice",
+			slice:     nil,
+			predicate: func(i int) bool { return i > 5 },
+			want:      nil,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := filter.Filter(tt.slice, tt.predicate)
+			require.Equal(t, tt.want, got, "Filter() = %v, want %v", got, tt.want)
+		})
+	}
+}
+
+func TestInPlace(t *testing.T) {
+	tests := []struct {
+		name      string
+		slice     []int
+		predicate filter.Predicate[int]
+		want      []int
+	}{
+		{
+			name:      "all match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 0 },
+			want:      []int{1, 2, 3, 4, 5},
+		},
+		{
+			name:      "some match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 3 },
+			want:      []int{4, 5},
+		},
+		{
+			name:      "none match",
+			slice:     []int{1, 2, 3, 4, 5},
+			predicate: func(i int) bool { return i > 5 },
+			want:      []int{},
+		},
+		{
+			name:      "empty slice",
+			slice:     []int{},
+			predicate: func(i int) bool { return i > 5 },
+			want:      []int{},
+		},
+		{
+			name:      "nil slice",
+			slice:     nil,
+			predicate: func(i int) bool { return i > 5 },
+			want:      nil,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := filter.InPlace(tt.slice, tt.predicate)
+			require.Equal(t, tt.want, got, "Filter() = %v, want %v", got, tt.want)
+		})
+	}
+}
